red team/menacraft-main/requirements.txt
pandas>=0.25.1
seaborn>=0.9.0
numpy>=1.17.0
scipy>=1.3.0
scikit-learn>=0.21.0
networkx>=2.3
requests>=2.25.0
beautifulsoup4>=4.9.0
python-dotenv>=0.19.0